refactor(iris): reorganize services/iris into chat/ context/ transport/
Sort the now-17 iris/ files into three concern-based subfolders so the
directory listing maps to architecture instead of alphabet:

- chat/        UI/webview routing (chat*, helpContent, messageUtils, irisWebSocketMessageHandler)
- context/     state storage (contextStore + the Task 1-4 splits, sessionManager, sessionSyncUtils)
- transport/   WS session infrastructure (irisWebSocketSessionClient)

iris/index.ts public surface is unchanged — every existing import via
'../services/iris' keeps working. Only one external file
(chatViewStatePresenter) used a deep path and was updated. Internal
imports between iris/ files were rewritten to the new relative paths.

Mechanical move via git mv (so renames are tracked); no behavior change.
