+++ extension/src/extension/services/iris/chat/helpContent.ts
