+++ extension/src/extension/services/iris/context/sessionSyncUtils.ts
@@ -1,9 +1,9 @@
-import { ArtemisApiService } from '../../api';
+import { ArtemisApiService } from '../../../api';
 import { ContextStore } from './contextStore';
-import { extractIrisMessageContent } from './messageUtils';
-import { logger, LogCategory } from '../loggingService';
-import type { ActiveContext, IrisChatSession, IrisChatMessage } from '../../types';
-import type { ExtensionToWebviewMessage } from '../../../shared/messageContracts';
+import { extractIrisMessageContent } from '../chat/messageUtils';
+import { logger, LogCategory } from '../../loggingService';
+import type { ActiveContext, IrisChatSession, IrisChatMessage } from '../../../types';
+import type { ExtensionToWebviewMessage } from '../../../../shared/messageContracts';
 
 /**
  * Shared dependency bag for Iris services.
