+++ extension/src/extension/services/iris/context/trackedItemRepository.ts
@@ -1,5 +1,5 @@
 import type { StoredState } from './contextStateTypes';
-import type { TrackedExercise, TrackedCourse, ContextSource } from '../../types';
+import type { TrackedExercise, TrackedCourse, ContextSource } from '../../../types';
 import {
     byPriorityThenRecency,
     byLastViewedDesc,
