+++ extension/src/extension/services/iris/index.ts
@@ -1,10 +1,10 @@
-export { ChatContextManager } from './chatContextManager';
-export { ChatDiagnosticsService } from './chatDiagnosticsService';
-export { ChatMessageService } from './chatMessageService';
-export { IrisChatSessionService } from './chatSessionService';
-export { IrisWebSocketSessionClient } from './irisWebSocketSessionClient';
-export { IrisWebSocketMessageHandler } from './irisWebSocketMessageHandler';
-export type { IrisServiceDeps } from './sessionSyncUtils';
-export { IRIS_CHAT_HELP_MARKDOWN } from './helpContent';
-export { ContextStore } from './contextStore';
-export type { ChatContextReason } from './chatContextManager';
+export { ChatContextManager } from './chat/chatContextManager';
+export { ChatDiagnosticsService } from './chat/chatDiagnosticsService';
+export { ChatMessageService } from './chat/chatMessageService';
+export { IrisChatSessionService } from './chat/chatSessionService';
+export { IrisWebSocketSessionClient } from './transport/irisWebSocketSessionClient';
+export { IrisWebSocketMessageHandler } from './chat/irisWebSocketMessageHandler';
+export type { IrisServiceDeps } from './context/sessionSyncUtils';
+export { IRIS_CHAT_HELP_MARKDOWN } from './chat/helpContent';
+export { ContextStore } from './context/contextStore';
+export type { ChatContextReason } from './chat/chatContextManager';
