+++ extension/test/unit/provider/contextStore.test.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert';
 import * as vscode from 'vscode';
-import { ContextStore } from '../../../src/extension/services/iris/contextStore';
+import { ContextStore } from '../../../src/extension/services/iris/context/contextStore';
 import { MockExtensionContext } from '../mocks/vscodeMocks';
 import { ActiveContext } from '../../../src/extension/types';
